+++ changelog.md
@@ -1,3 +1,7 @@
+## [0.0.2] - 2022-22-02 - R.S.
+### Ajout
+-Mise en place de la réponse au message.
+
 ## [0.0.1] - 2022-21-02 - R.S.
 ### Ajout
--API Steam pour Unity
+-API Steam pour Unity.